+++ addOns/ascanrules/src/test/java/org/zaproxy/zap/extension/ascanrules/PersistentXssScanRuleUnitTest.java
@@ -48,14 +48,12 @@ class PersistentXssScanRuleUnitTest extends ActiveScannerTest<PersistentXssScanR
         assertThat(
                 tags.containsKey(CommonAlertTag.OWASP_2021_A03_INJECTION.getTag()),
                 is(equalTo(true)));
-        assertThat(
-                tags.containsKey(CommonAlertTag.OWASP_2017_A01_INJECTION.getTag()),
-                is(equalTo(true)));
+        assertThat(tags.containsKey(CommonAlertTag.OWASP_2017_A07_XSS.getTag()), is(equalTo(true)));
         assertThat(
                 tags.get(CommonAlertTag.OWASP_2021_A03_INJECTION.getTag()),
                 is(equalTo(CommonAlertTag.OWASP_2021_A03_INJECTION.getValue())));
         assertThat(
-                tags.get(CommonAlertTag.OWASP_2017_A01_INJECTION.getTag()),
-                is(equalTo(CommonAlertTag.OWASP_2017_A01_INJECTION.getValue())));
+                tags.get(CommonAlertTag.OWASP_2017_A07_XSS.getTag()),
+                is(equalTo(CommonAlertTag.OWASP_2017_A07_XSS.getValue())));
     }
 }